Null check on video file
git-svn-id: https://alma.andrew.cmu.edu:8443/svn/alice3/trunk@947 c85dc8a7-84c8-304b-b635-f907e77af1ba

+++ x_youtube/src/org/alice/media/VideoCapturePane.java
@@ -876,7 +876,10 @@ public abstract class VideoCapturePane extends LineAxisPane implements ActionLis
 
 	private void close()
 	{
-		this.recordedMovieFile.delete();
+		if (this.recordedMovieFile != null)
+		{
+			this.recordedMovieFile.delete();
+		}
 		this.onClose();
 	}
